# server/controllers/CadenceRT_db/generated/GeocodeControllerGenerated.js
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE GeocodeControllerGenerated.js PLEASE EDIT ../GeocodeController.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
import Properties from "../../../properties";
import GeocodeModel from "../../../models/CadenceRT_db/GeocodeModel";
import ErrorManager from "../../../classes/ErrorManager";
import { authorize } from "../../../security/SecurityManager";
import GeocodeController from "../GeocodeController";

const generatedControllers = {
  /**
   * Init routes
   */
  init: router => {
    const baseUrl = `${Properties.api}/my/geocode`;
    router.post(baseUrl + "", authorize([]), GeocodeController.create);
    router.delete(baseUrl + "/:id", authorize(["SUPERADMIN"]), GeocodeController.delete);
    router.get(baseUrl + "/findByclass/:key", authorize([]), GeocodeController.findByclass);
    router.get(baseUrl + "/findByr_owner/:key", authorize([]), GeocodeController.findByr_owner);
    router.get(baseUrl + "/:id", authorize([]), GeocodeController.get);
    router.get(baseUrl + "", authorize([]), GeocodeController.list);
    router.post(baseUrl + "/:id", authorize([]), GeocodeController.update);
  },


  // CRUD METHODS


  /**
  * GeocodeModel.create
  *   @description create point set
  *   @param String class
  *
  */
  create: async (req, res) => {
    try {
      const result = await GeocodeModel.create(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * GeocodeModel.delete
  *   @description CRUD ACTION delete
  *   @param String id Id
  *
  */
  delete: async (req, res) => {
    try {
      const result = await GeocodeModel.delete(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * GeocodeModel.findByclass
  *   @description CRUD ACTION findByclass
  *   @param Objectid key Id of the resource class to search
  *
  */
  findByclass: async (req, res) => {
    try {
      const result = await GeocodeModel.findByclass(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * GeocodeModel.findByr_owner
  *   @description Search by findByr_owner
  *   @param String key Id of the resource r_owner to search
  *
  */
  findByr_owner: async (req, res) => {
    try {
      const result = await GeocodeModel.findByr_owner(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * GeocodeModel.get
  *   @description CRUD ACTION get
  *   @param ObjectId id Id resource
  *
  */
  get: async (req, res) => {
    try {
      const result = await GeocodeModel.get(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * GeocodeModel.list
  *   @description CRUD ACTION list
  *   @param String class
  *
  */
  list: async (req, res) => {
    try {
      const result = await GeocodeModel.list();
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  /**
  * GeocodeModel.update
  *   @description CRUD ACTION update
  *   @param ObjectId id Id
  *
  */
  update: async (req, res) => {
    try {
      const result = await GeocodeModel.update(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  // Custom APIs

};

export default generatedControllers;