# server/swagger.yaml
openapi: 3.0.0
info:
  title: CadenceRT
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /my/geocode/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Geocode
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Geocode
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Geocode"
    post:
      summary: CRUD ACTION update
      tags:
        - Geocode
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Geocode"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Geocode"
  /my/geocode/findByclass/{key}:
    get:
      summary: CRUD ACTION findByclass
      tags:
        - Geocode
      parameters:
        - name: key
          in: path
          description: Id of the resource class to search
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Geocode"
  /my/geocode/findByr_owner/{key}:
    get:
      summary: Search by findByr_owner
      tags:
        - Geocode
      parameters:
        - name: key
          in: path
          description: Id of the resource r_owner to search
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Geocode"
    
  #-----  Stakeholder  -----
...
  /my/advanced/treedata/join-to-node/{handle}:
    patch:
      summary: Add a new member to a tree node
      tags:
        - Treedata
      parameters:
        - name: class
          in: path
          description: Optional
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
...
  /my/advanced/treedata/{id}:
    delete:
      summary: delete by name
      tags:
        - Treedata
      parameters:
        - name: id
          in: path
          description: Id Treedata
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: Get by Id
      tags:
        - Treedata
      parameters:
        - name: id
          in: path
          description: Id Treedata
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Treedata"
    post:
      summary: CRUD ACTION update
      tags:
        - Treedata
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Treedata"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Treedata"
  /my/advanced/treedata/findByclass/{key}:
    get:
      summary: By Class
      tags:
        - Treedata
      parameters:
        - name: key
          in: path
          description: Id of the resource class to search
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Treedata"
  /my/advanced/treedata/findByref_inlaw/{key}:
    get:
      summary: Search findByref_inlaw
      tags:
        - Treedata
      parameters:
        - name: key
          in: path
          description: Id of the resource ref_inlaw to search
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Treedata"
  /my/advanced/treedata/findBys_lastchild/{key}:
    get:
      summary: CRUD ACTION findBys_lastchild
      tags:
        - Treedata
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Treedata"
  /my/advanced/treedata/findBys_parent/{key}:
    get:
      summary: By parent
      tags:
        - Treedata
      parameters:
        - name: key
          in: path
          description: Id of the resource s_parent to search
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Treedata"
  /my/advanced/treedata/{id}/gets_parent:
    get:
      summary: gets_parent / members
      tags:
        - Treedata
      parameters:
        - name: id
          in: path
          description: ID of Treedata from s_parent
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    
  #-----  User  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    Geocode:
      properties:
        _id:
          type: string
        _created:
          type: integer
        class:
          type: string
        o:
          type: number
        series:
          type: boolean
        x:
          type: number
        xy:
          type: number
        y:
          type: number
      required:
        -  _id
        - _created
        - class
        - series
        - x
        - y
...
    Treedata:
      properties:
        _id:
          type: string
        _created:
          type: integer
        _modified:
          type: integer
        class:
          type: string
        description:
          type: string
        handle:
          type: string
        metadata:
          type: string
        orphan:
          type: boolean
        ref_inlaw:
          type: string
        ref_last:
          type: string
        ref_parent:
          type: string
        ref_twin:
          type: string
        status:
          type: boolean
      required:
        -  _id
        - class
        - handle
...